Add packaging support.
This includes 3 packaging layouts. The development option is useful for
compiling all the dependencies into a single directory so wrapper scripts can be
made to work easily in-tree. The standalone mode builds an uber jar for easy
deployment without an additional packaging step. And the package mode builds a
directory for deb/rpm packages, excluding dependencies that should be provided
by other debs/rpms.