Fix consumer worker wait timeout so it is never negative.
By the time the timeout is computed, the next timeout may have expired and some
Java implementations don't handle negative values for Object.wait. Additionally,
we need to carefully handle a timeout of zero since that has special meaning for
Object.wait. Fixing this also revealed an incorrect timeout computed in the
MockConsumerQueue, which should never be larger than the caller-specified
timeout.