Improve scheduling of consumer worker threads.
Fix the ConsumerWorker threads to wait on read tasks using a priority queue to
determine the thread that needs to do work soonest when no threads have active
work. When computing the next time, consider both the current backoff and the
expiration of the task so that expiring tasks don't need to wait for a backoff
period to complete before completing.