Add Avro consumer support.
The type of the consumer is selected during consumer creation and the response
entity content types for consumer read requests is restricted to the type
specified at creation. If unspecified, the default is binary. This patch adds
AvroConverter support for Avro -> JSON which computes approximate data sizes as
it performs the conversion in order to limit total response size.