Improve handling of missing topics and producer exceptions, and return per-message information on produce requests.
This makes three major changes:
1. Properly handle producer exceptions, returning per-message error information
back to the user in the ProduceResponse. We consider the request successful as
long as the only error we hit is returned by the producer future, in which case
we include that information in the response.
2. Missing topics are now handled properly for either broker setting of
auto.create.topics.enable. When auto creation is disabled, Kafka currently
generates a timeout exception, which is now handled properly and passed back to
the user with an error code. More generally, all Kafka exceptions are converted
into an error code indiating either retriable or non-retriable errors, or
triggers a 500 for non-Kafka exceptions.
3. As a side effect of these changes, this fixes the data returned on produce
requests. It now includes per-message data. Originally the responses only
included aggregate information about the last per-partition offset seen in the
published messages. However, properly handling errors requires per-message
notifications of success. To handle this, each response includes either
partition and offset or error_code and error message.

This fixes #24. It also fixes the producer half of #29.