Add support for producer and consumer configuration overrides.
Pass configuration options provided to the KafkaRestConfig along to producer and
consumer instances. This allows the operator of the proxy to provide overrides
for the default configs. This is critical for some site-wide config options like
compression that should not be specified per-request.

This approach has the drawback that switching producer/consumer implementations
potentially causes breakage since config options may change. This is
particularly a potential issue when we upgrade from the old consumer to the new
consumer in the near future. This may cause some additional complexity for
deployment, but since extra config options shouldn't cause a problem, there
should be a safe upgrade path where both sets of options are provided. This also
doesn't yet expose the full set of options to user requests, so it's less of a
commitment in that it affects operators of the proxy, but not users of the API.

Fixes #51.