Handle conflicting IDs and separate IDs used in the REST proxy and by Kafka's consumer implementation.
Add a check for duplicate IDs before creating a consumer and adding it to the
map in ConsumerManager. Also add a "name" field to ConsumerInstanceConfig to
allow control over the naming of the instance config within the single REST
proxy without setting the ID of the underlying Kafka consumer, since this can be
useful if a limited set of consumer instances within a group is desired. Update
the documentation to mark the "id" field as deprecated since it probably doesn't
provide the behavior users expect and without this fix it can cause problems if
the user ends up with duplicates. Kafka doesn't expect it to be used this way
except for testing anyway. Also add the name field and update all the example
cURL requests to use the recommended default where both "id" and "name" are
omitted, and consumers are deleted properly when they are no longer needed.

Fixes #65.