Bug fix to consumer request max bytes
The logic for max bytes doesn't work.  The loop that processes messages
correctly breaks out when a message takes bytes over the max bytes but
then the final check fails to account for the in-progress message and
doesn't terminate the connection.

Have suspended this work due to a flaw in the tests that needs fixing
before making changes to the core code.