Fix to Callback Tests
Current callback tests rely on asserts within the callbacks.  It appears
that these asserts raise exceptions that are then handled thus
smothering the assertion failure.o

Have moved the asserts externally to the callback.

To demonstrate the problem the second test in testConsumerNormalOps has
an assertion that clearly fails and yet the test does not fail.