Fix to correct max bytes bug along with test fixes
There is a bug in the ConsumerReadTask in the handling of max bytes.
The code stops processing messages once the max bytes is exceeded but
will not return until timeout elapses.

In fixing the above I have also had to fix the tests.  Currently a
MockTime class is used.  The aim of this class is to allow precise
checks against timing however the ConsumerReadTask does not fit the
pattern that the mock assumes (the mock assumes use of sleep and or
onWait but the ConsumerReadTask does not use either of these).