CLIENTS-304: Fix AvroConsumerRecord constructors and handling of topic for v1 consumer API.
AvroConsumerRecord had two constructors with @JsonProperty parameters which caused it to sometimes fail to properly
deserialize. Additionally, adding the topic field to ConsumerRecord for the v2 API resulted in null entries in the v1
API. This is confusing; instead, we should be filling that in since it is a backwards compatible addition and keeps the
output uniform.