CPKAFKA-1865:  Performance degrades drastically after running for extended periods (#504)
The problem stemmed from the fact that the proxy kept a priority queue of the consumers, ordered by their expiry. Whenever we wanted to update a consumer's expiry, we would re-enter it into the structure so that a re-ordering happens. There was a bug where when re-adding a consumer into the structure, we wouldn't remove its previous instance. This caused the structure to grow very large, causing O(N) operations such as consumer delete to block the proxy for long periods.

We now remove the consumersByExpiry data structure totally since it is not needed.
The ExpirationThreads now traverse the whole consumers structure and remove any expired consumers from there. Chances are that the overall number of consumers will be very low and that would not cause any performance impact