CPKAFKA-1681:  Add long polling functionality (#481)
This commit enables long-polling in V1 and V2 APIs of the REST proxy.

Adds a new config setting:
fetch.min.bytes - the amount of bytes of records the proxy will collect before immediately returning a response. It is configurable per consumer.
This new setting controls how the user->proxy connection works, not the underlying consumer->broker.

Previously, consumer.request.timeout.ms would control when the proxy sent back a response to the user. The default value was 1000ms but setting it explicitly to any value would cause it to be passed to the underlying consumer connection. This caused a NullPointerException when set incorrectly, due to the fact that configuration exceptions were ignored in the V2 API. The `consumer.request.timeout.ms` config is no longer passed to the consumer but instead set to 30s.


The proxy now raises exceptions and returns proper 4xx response when invalid V2 Consumer settings are given.