Split the whole ProduceRequest/ProduceRecord type hierarchy into V1 and V2 serialization classes.
This looks like a pretty scary change, but it is actually pretty safe. I tested/verified the changes in code first, without any change to the tests, before changing the tests to use the new types. All requests still accept the same possible inputs, I just split the classes.

The changes can be summarized as:

* Created `AvroPartitionProduceRequest`, `AvroTopicProduceRequest`, `BinaryPartitionProduceRequest`, `BinaryTopicProduceRequest`, `JsonPartitionProduceRequest` and `JsonTopicProduceRequest`, for V1 produce endpoints.
* Created `BinaryPartitionProduceRequest`, `BinaryTopicProduceRequest`, `JsonPartitionProduceRequest`, `JsonTopicProduceRequest`, `SchemaPartitionProduceRequest`, `SchemaTopicProduceRequest` for V2 produce endpoints.
* Created (new) `ProduceRequest` and (new) `ProduceRecord` helper classes.
* Removed entire (old) `ProduceRequest` type hierarchy.