CLIENTS-1499: Add LIST/GET Partitions V3 endpoints. (#635)
This PR adds endpoints to list and get partitions:

* /v3/clusters/{clusterId}/topics/{topicName}/partitions
* /v3/clusters/{clusterId}/topics/{topicName}/partitions/{partitionId}

A non-inclusion was (beginning|end)_(offset|timestamp). Somehow, Admin.listOffsets does not seem to be working as I would expect. It always returns beginning_offset as 0 (regardless of the partition being empty or not), and all timestamps are (EPOCH - 1ms). We can add this information later.