CLIENTS-1514: Defaults to JSON:API Content-Type when no Accept header in request. (#643)
Right now, for requests that return 204 No Content on successful completion, if you don't specify a Accept header, and and error happens, 500 is returned, instead of the code mapped to the error. That is because the server is defaulting to `text/html` Content-Type, and Jackson does not know how to serialize the JSON error response body to `text/html`. This PR makes sure the Content-Type of responses is `application/vnd.api+json` by default.