Remove all methods from AdminClientWrapper. (#658)
Produce to-topic and to-partition have been split out in a separate "resource", so that TopicsResource and PartitionsResource can no longer depend on KafkaRestContext.

Note: Unfortunately, because the way Jersey request dispatching works, we cannot have a single "ProducersResource", with all the produce methods. Jersey does not deal very well with ambiguous path prefix matching.

AdminClientWrapper will be removed in a separate PR (need to update security plugins first).