Fix REST Proxy injection-based authentication. (#672)
Background: Confluent security plugins for REST Proxy perform authentication in a Post-Match Filter. Upon successful authentication, it sets up AdminClient connection accordingly. Later on, we bind Admin to this AdminClient instance created during authentication, and this instance is what is injected in the controllers/resources.

The problem: JSR 339/JAX-RS 2.0 specification, Section 3.7.2, Step 3 [1], specifies that the resource instance must be created in the Resource Match phase, which happens before the Post-Match Filter phase (see Appendix C). Therefore, at the time the Admin is injected in the resources, AdminClient has not been properly set yet. A default unauthenticated instance is injected instead.

The solution: A possible solution would be to move the authentication to the Pre-Match Filter phase. Unfortunately that is undesirable for a few reasons (for example, without matched resource information, it can't properly negotiate the content-type of the response in case of an authentication error). The solution implemented here instead is to inject a Provider of Admin, instead of Admin itself. By the time we call `get()` in the provider (Method Invocation phase, see Appendix C), authentication has already happened, and the correct AdminClient instance is acquired.

[1] https://download.oracle.com/otn-pub/jcp/jaxrs-2_0_rev_A-mrel-spec/jsr339-jaxrs-2.0-final-spec.pdf