Migrate off JSON:API. (#686)
After careful consideration, we have decided that the format introduced in the CP 5.5 preview of the V3 API, JSON:API, is not as user friendly as we would like that API to be. We have decided to go back to a proprietary format for our JSON APIs. This PR is part 1 of N on getting off JSON:API. It converts all APIs request/response bodies to the new format.