KC-613: Split share/java/kafka-rest in kafka-rest-bin and kafka-rest-lib. (#737)
This PR splits the packaged contents of kafka-rest in two folders share/java/kafka-rest-bin and share/java/kafka-rest-lib. The later is to be used when including kafka-rest as library in another artifact, while both should be used when running kafka-rest as a standalone service.

This PR also changes the test environments to use the latest packaged cp-kafka-rest image as base image. This is to make sure the changes in this PR work with cp-kafka-rest image.