KREST-26: Add a blocklist config to allow disabling individual endpoints. (#756)
Adds a config api.endpoints.blocklist to allow disabling individual endpoints. Each resource class/method is now annotated with a @ResourceName annotation that uniquely identifies a endpoints or set of endpoints. The config receives a list of such names to disabled.

If an endpoint is disabled, it will return HTTP 404 Not Found for GET requests, and HTTP 405 Method Not Allowed for everything else.