KREST-147: Make absolute URL generation consistence in V2 and V3.
Config advertised.listeners was added in 6.0.0 to specify which base path to use when generation absolute URLs in V3 API responses. Consume V2 API has a similar mechanism, that uses the host.name and port configs. This PR unifies both mechanisms, so that the URLs generated in both V2 and V3 are based on the same base path.