Remove kafka-rest-console-scripts.
Reason: We don't need a separate target for scripts, specially because the only scripts we have are to start and stop the server, which are useless without the rest of the server's artifacts.

Real reason: I have no idea what this is needed for, and I mostly want to see what breaks (if anything) if I delete it.