Multiplex all produce requests into a single Producer. (#787)
Background: Currently the way V2 produce works internally is that REST Proxy keeps one producer for each serialization format (`BINARY`, `JSON`, `AVRO`, `JSONSCHEMA`, `PROTOBUF`). Based on the format of the incoming request, it will decide to which producer to send the request to.

Problem: This works "well" for V2 because in V2, both key and value must use the same serialization format. I say "well" in quotes because it still requires maintaining 5 producers up, but that is still manageable. In V3, unfortunately, that won't do. One of the design differences of produce V3 is that it decouples the serialization format of key and values. If we were to follow the existing design, we would need to keep 25 (5x5) producers, which is not managleable.

Solution: This PR changes the produce internals to use a single producer instead, a byte-array producer. To achieve it, we change REST Proxy to serialize the records BEFORE sending them to the producer, instead of handing the producer the unserialized record for the producer itself to serialize.

Detailed Solution:

The existing flow looks something like:
* `schema = getSchema(request) // this is only done for AVRO, JSONSCHEMA and PROTOBUF`
* `producer = getProducer(request)`
* `producer.producer(schema, request)`

The new flow looks like:
* `schema = getSchema(request) // this is only done for AVRO, JSONSCHEMA and PROTOBUF`
* `serialized = serializeRecord(schema, request)`
* `binaryProducer.produce(serialized)`

Step (1) is handled by `SchemaManager`, step (2) is handled by `RecordSerializer`, and step (3) is handled by `ProduceController`.

Testing: Great care has been taken to achieve exact same behaviour, since we are basically reimplementing the entire V2 produce API. All the integration tests (produce is arguably the most tested feature on REST Proxy) are largely unchanged. `SchemaManager`, `RecordSerializer` and `ProduceController` are throrougly unit-tested for parity with the now deleted `ProducerPoll`, `NoSchemaRestProducer` and `SchemaRestProducer`.