KREST-600: Fix KafkaConsumerManagerTest.testBackoffMsControlsPollCalls flaky test. (#838)
Current way it is being tested, it relies on Thread.sleep to try to peek into the KafkaConsumerManager at right time. Problem is, 1. it is using the wrong time (it assumes the backoff is 250ms, when actually it is 50ms), and 2. it is peeking the wrong thing. I just looks if there's a delayed task, but that gives no guarantees the task is being retried at the given backoff. So I rewrote it to actually the backoff.

It was previously consistently failing 5% of time (out of 1000 runs). Now it fails 0%. :)

Example failure: https://jenkins.confluent.io/job/confluentinc/job/kafka-rest/job/6.0.x/480/testReport/junit/io.confluent.kafkarest.v2/KafkaConsumerManagerTest/testBackoffMsControlsPollCalls/