Always allow OPTIONS requests. (#862) (#863)
Because OPTIONS requests are handled by the default jersey
OptionsMethodProcessor, the allowlist feature does not play nice with
it, as it will block anything that is not annotated with something
that is in the allowlist, and the jersey processor is clearly not. We
also have the problem that this processor handles OPTIONS requests for
all endpoints, and we don't have a good way on intercepting it making
it work nice with the allowlist.

The current stop gap solution is to allow OPTIONS for all
endpoints. We can investigate better solutions afterwards.