Use a more lenient test for testBackoffMsControlsPollCalls.
Previously on #838, we tried to fix testBackoffMsControlsPollCalls,
but we inadvertently made the test too strict. It would assume a
constant number of 2 + (timeoutMillis / backoffMillis) poll calls,
uniformly distributed across timeoutMillis. It would allow only for a
small deviation of -1%/+10% of the expected time for the poll
calls. It turns out these assumptions are too strict and test
consitently fails during PR test runs, and sometimes on branch test
runs.

The new strategy implemented in this PR instead tests that there's no
window of size backoffMillis with more than 2 poll calls, and no
window of size 2 * backoffMillis with no poll calls. This indirectly
tests that the poll calls are distributed across timeoutMillis,
without expecting too much about the actual timing of each poll call.