Pass KafkaRestContext to HK2 via ContainerRequestContext.
Currently the mechanism used by the various request
processing stages (request/response filters/interceptors and request
handlers) to communicate with each other is
KafkaRestContextProvider. This class owns a thread-local, which is
supposed to keep state per-request-handling-thread (although that's
not 100% true in the presence of async requests). If this thread-local
is not set, than Kafka REST fallsback to a default KafkaRestContext,
which is kept in the KafkaRestContextProvider in a static variable, to
preserve singleton property. This has worked fine so far because up to
this point there has always been a single KafkaRestApplication running
per JVM, but that assumption is not valid anymore. Since the static
cannot isolation the singleton instances accross multiple
KafkaRestApplication instances in the same JVM, another approach is
taken here.

What this PR does is to use instead the ContainerRequestContext to
keep the state per-request state. This approach has the nice benefit
of working, regardless of the request being suspended or not. The
default singleton KafkaRestContext is owned by HK2 now, and will be
isolated per injector (which will be one per application).

A follow-up PR will be made to delete KafkaRestContextProvider, once
all usages have been removed.