KREST-1673 Add error responses to the OpenAPI spec.
This adds some common error responses to all endpoints in our
OpenAPI spec. Unfortunately OpenAPI does not provide a nice way
to handle these in a single place, so there's a lot of duplication.
- See https://github.com/OAI/OpenAPI-Specification/issues/398,
https://github.com/OAI/OpenAPI-Specification/issues/521,
https://github.com/OAI/OpenAPI-Specification/issues/566.

The error responses also include some examples that will hopefully
be useful for users and developers. However the examples are by no
means exhaustive.