Fix NPE causing most integration tests to fail
Recent AK changes related to the removal of ZooKeeper have
necessitated the accompanying changes in kafka-rest's tests.
https://github.com/confluentinc/kafka-rest/pull/893 has done
that, but unfortunately it also removed the initialization of
the zkConnect field which then caused an NPE when it ended
up being added to a ConcurrentHashMap as a null.

This attempts to fix the problem by the simplest of approaches:
putting back the initialization of zkConnect.