KREST-2260: Use super user when creating topic in AuthorizationErrorTest.
Prior to https://github.com/confluentinc/kafka-rest/pull/893, this topic was being created directly in Zookeeper, bypassing Kafka authorization entirely.

Now that we are using Kafka to create the topic, we need to pass through authorization. This changes uses the super-user to create the topic, so that we don't fail authorization.