KREST-1967 Address review comments
The rate limit needs to be global, not per connection, so I have made the rate limit tracker static, as well as the grace period tracker.

Missing a space in the TooManyRequestsException.

Added a warn for when the grace period is exceeded so there is information about when this is happening.