KREST-1967
Address review comments - the rateCounter is now injected into the contructor

I also spotted a bug while testing.  If the MappingIterator was not also closed, an existing connection could keep spamming the produce via the strem, because the sink.close wasn't actually closing the connection.

This meant that no new connections could be made, because the rate limit continued to be exceeded.