KREST-1936 Remove mask filtering of producer configs.
This change removes the filtering of all properties not explicitly
known to be valid for producer configs. This has demonstrated to be
a problem for producer configs before (see KREST-481), and now causes
a problem for other interceptor configs.

Note that such filtering doesn't work in the general case in the
face of extensibility mechanisms that Kafka has (like interceptors)
that don't provide a way to register their expected and allowed config
in advance.

That same filtering has been left unchanged for the different SR-related
configs, despite a history of causing problems there too (see KREST-1187).
Removing that same filtering for SR-related configs requires more
refactoring so it won't be addressed in this change.