Bind NoSchemaRecordSerializer as Singleton. (#933)
This is to avoid creating a new NoSchemaRecordserializer for each record in a stream. Normally, this instances should be very light and short-lived, so instantiating them should not be a problem. Unfortunately, a new `KafkaJsonSerializerConfig` is created for each `NoSchemaRecordSerializer`, and that thing is everything but light. In special, `AbstractConfig` (super-type) has the bad habit of logging all configs at creation, and when you that doing thousands of times per second, that usually takes a lot of cpu.