KREST-3430 Optimize rate-limiting exceptions
The StatusCodeExceptions that we throw in great numbers when
rate-limiting can take a significant amount of time for filling
their stack traces (up to 1% of a flamegraph taken during
rate-limiting). These stack traces however are not really
interesting for us, as the exceptions themselves always
originate from our rate-limiting infrastructure.

The change overrides fillInStackTrace() instead of utilizing
the RuntimeException ctor that takes a writableStackTrace
flag, because while slightly more efficient, the latter
approach is much more cumbersome.