Close produce stream on idle time and unauthenticated.
When handling produce requests, the request handling thread is held for as long as the input stream is open. Seeing we use a fixed size thread pool for request handling, a bad user could potentially deny any request from being handled by starving all request handling threads with open produce streams. That's attack is made worse by:

 1) we don't close the stream on unauthenticated from Kafka, so the user doesn't even need to have valid credentials
 2) we don't close the stream on idle time, so the user doesn't even need to send any data

This PR closes the two holes mentioned above. The server will now close the stream both if it hasn't seen requests in the stream within some or if any request produces an unauthenticated response from Kafka. Effectively, what that achieves is that we are forcing now each stream to authenticate with Kafka within some predefined time.

You can configure the idle time for the stream using the property `streaming.response.max.idle.time.ms` (default to 30s). Notice that there's an initial negotiation step where the stream is open but no bytes have actually been sent over the stream. In this case, the idle time is controlled not by Kafka REST, by Jetty, which already defaults it to 30s. A follow up PR will be done in `rest-utils` to add a way to configure this other Jetty idle time. It's strongly recomended that both idle times to be configured to the same value.