KREST-4113 Don't delete all ACLs if an empty DELETE request is received.
This change prevents the simplest DELETE request (a no-parameter one)
from deleting all ACLs. This is done by making the default values
for the non-nullable query params invalid. This forces the explicit
setting of these query params by the user. The previous default
values have not been documented which, in addition to the note
below about passing some parameters only not working, is why we
are not concerned about breaking backward compatibility.
- In addition to that, some inconsistencies in the delete request
  validation have been addressed.

For some reason, requests with partially filled query params are
not currently hitting the "Delete ACL" endpoint - it should be
investigated whether that's due to a testing problem, Jersey bug,
or something else.