KREST-3322 Remove grace period and swap out custom rate limiter (#961)
* KREST-3322 First pass of removing grace period

This PR removes the grace period that was originally added for rate limiting REST produce requests, and also replaces the custom rate limiter with the Guava/Resiliance4J limiter that is currently used to limit admin requests.

Although grace period was originally included in the 1 pager for produce, with more experience of rate limiting we have determined that using a simple library limiter is a better solution.

When a streamed produce is limited, the connection is no longer closed, instead a 429 is returned and the message is not processed.

The Guava/Resiliance4J/Null high level request rate limiters are injected, with configuration taken from the overall rate limits.

For produce, the produce specific limits need to be used, so I have made new Factory classes for both bytes and count limiters (they take different values) that extend the original factory (as otherwise I can’t override the configuration values).

The new factories call super on the original RequestRateLimiterFactory to initialise the RequestRateLimiter, and then store a local copy of the RequestRateLimiter to be used for the actual rate limiting. Unfortunately we need to store a copy of the RequestRateLimiter, as I need to be able to inject either a produce bytes limiter or a produce count limiter but the new limiter classes can’t extend all of the Guava, Resiliance4J and Null Rate limiters (because Java). I had to make the new injectable classes just extend the RequestRateLimiter and override the rateLimit method to call the underlying saved RateLimiters limit method. Convoluted, yes. Sorry. Open to any suggestions to make it simpler!

The cost of a counter based limit is 1, the cost of a bytes based limit is the size of the message in bytes.

Produce action had some code that I didn’t understand (and I think was wrong), where it didn’t rate limit at all if there was no message size available - I’ve taken that out so that rate limiting always happens, but if there is no message size, the bytes based limiting is not done. Looking through the code, the message size is always there (added in the deserialise method of the ProduceRequestSerdeModule).

The alternative is the we barf if the size is missing - what do you think?

In the ProduceRateLimiters, I quick return if rate limiting is disabled - this feels more efficient than allowing it to fall through, for both bytes and count, to the null limiter to return on.

I’ve taken out a number of the ProduceActionTest unit tests because they were actually testing the rate limiter, not produce action, and that’s either now done in ProduceRateLimiters or is unnecessarily checking the behaviour of the external libraries.

* KREST-3322 Address review comments

Conflicts:
kafka-rest/src/test/java/io/confluent/kafkarest/resources/ProduceRateLimitersTest.java