Fix context initialization in ConsumersResource.
Recent muckrake failures pointed to problems in the
initialization of our Consume v2 resource on 7.1.x and
above. It turned out that with the recent switch to
fully handling the KafkaRestContext lifetime via our DI
mechanisms, necessary context initialization happening
in filters is done after the eager context initialization
of KafkaRestContext for ConsumersResource. That lead to
NPEs and HTTP 500s for calls to the Consume v2 API.

The fix here is to switch to lazy context initialization
via javax.inject.Provider. This ensures that the necessary
post-matching request filters have been executed by the
time the context actually is needed.