KREST-4591 add topic recreate logic to cluster test harness
Most of the test flakes in this class are due to failures in the setup and the topics not being created.

I can't reproduce this locally unfortunately, but I have put in a pile of retry logic to allow more time, and then have another try at creating the environment.

Also add in a test retry for a flakey test.