KREST-4605 Deflake all the tests that extend AbstractConsumerTest
KafkaConsumerReadTask
exception stack trace not logged, added in so we have a starting point for debugging the intermittently failing test

AbstractConsumerTest
a 404 is returned somewhere in startConsumerMessages (can't prove where, but one of the assertOKResponses fails with a 404) so check for the 404 on both, and then retry after a pause if it is returned.  Fail the second time
The number of messages consumed doesn't always match how many we expect.  I could reproduce this locally, and a single retry didn't always help, so I've added in two.  If it still fails we can look at reworking this a different way
consumeForTimeout is inherently flakey as it is so reliant on times.  I've doubled the wait period slack to 1s, from 501ms to see if that helps at all, but otherwise we will need to look at a mocked back end.

ClusterTestHarness
Made a generic produce method, as this was repeated with different object types
The produce was also failing in some cases so I've added some retry logic to try and resend each message 3 times with a pause each time before failure